Release/first release

+++ Api1/application/FileService.py
@@ -1,11 +1,11 @@
 from Configs.app import app
-from typing import Dict, BinaryIO
 from werkzeug.utils import secure_filename
 import os
 from exceptions.UploadedFileException import UploadedFileException
 from werkzeug import FileStorage
 import pathlib
 import ntpath
+from Aop.loggingDecorator import loggingDecorator
 
 
 class FileService(object):
@@ -15,21 +15,28 @@ class FileService(object):
     def __init__(self):
         pass
 
+    @loggingDecorator()
     def saveImageFileToDir(self, file: FileStorage, userId: str) -> str:
         return self._saveOrUpdateImageToDir(file, userId)
 
-    def deleteImageFile(self, userId: str, originalFilePath: str) -> bool:
+    @loggingDecorator()
+    def deleteImageFile(self, userId: str, originalFilePath: str = None) -> bool:
         # extract file name from path
-        originalFileName = self._extractFileName(originalFilePath)
-        # remove existing image if isUpdate is True
-        os.remove(os.path.join(app.config['UPLOAD_FOLDER'], str(userId), originalFileName))
+        # add this condition because selenium testing sometimes cause delete image request before add image request completed
+        if originalFilePath is not None:
+            originalFileName = self._extractFileName(originalFilePath)
+            # remove existing image if isUpdate is True
+            os.remove(os.path.join(app.config['UPLOAD_FOLDER'], str(userId), originalFileName))
 
+    @loggingDecorator()
     def _allowed_file(self, filename: str) -> bool:
         return '.' in filename and filename.rsplit('.', 1)[1].lower() in self._allowedExtension
 
+    @loggingDecorator()
     def _extractFileName(self, originalFilePath: str) -> str:
         return ntpath.basename(originalFilePath)
 
+    @loggingDecorator()
     def _saveOrUpdateImageToDir(self, file: FileStorage, userId: str):
 
         if file.stream is not None and self._allowed_file(file.filename):
